BackEnd/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib
python-dotenv
pydantic
email-validator
